fix: model setting dialog
